Introduce Standard ESM
This change enables a smooth transition between Node and ES module
formats and compensates for CJS and ESM incompatibility.

The tool used to enable this transition is
[esm](https://github.com/standard-things/esm).

The `esm` loader is intended for runtime use, so no build step or
pre-publish step is required of the user. There is a 1 time transform
cost per ESM module, after that the result is cached and loading is on
par with CJS. The loader itself when published will be a small, ~40kb,
zero dependency package.

Changes:

+ add `esm` as a dependency
+ add `esm` config to `package.json`
  + `"cjs": true` means we get to use `require` in files until we’re ready to convert to `import`
+ repurpose `lib/cli/index.js` to use `esm` loader
+ move `lib/cli/index.js` code to `lib/cli/main.js` and replaced `require`s with `import`s
+ remove `use strict;` from files as `esm` adds it by default
+ add `overrides` to `.eslintrc.js` to hint different parse option
+ add a note on ESM support

Previously, we ran into an issue w/ having to change our public API
`require`s:

```javascript
// Now
const EmberApp = require('ember-cli/lib/broccoli/ember-app');

// After conversion (w/o making changes)
const EmberApp = require('ember-cli/lib/broccoli/ember-app').default;
```

We don't want to break our public API of course, so we can do the
following:

```javascript
// lib/broccoli/ember-app/index.js
require = require('esm')(module)
module.exports = require('./main.js').default

// lib/broccoli/ember-app/main.js
/* current `ember-app.js` converted to ES module */
```

That way public API won't break.

Useful links:

+ [Prior art](https://github.com/ember-cli/ember-cli/pull/7331)
