+++ lib/utilities/require-as-hash.js
@@ -1,5 +1,4 @@
-'use strict';
-const stringUtils = require('ember-cli-string-utils');
+import stringUtils from 'ember-cli-string-utils';
 
 // Gathers subclasses of a certain specified base class into a hash.
 //
@@ -13,12 +12,12 @@ const stringUtils = require('ember-cli-string-utils');
 //   Hamster: Hamster // Same as require('./hamster.js')
 // }
 
+import { sync as globSync } from 'glob';
 
-const globSync = require('glob').sync;
-const path = require('path');
-const getCallerFile = require('get-caller-file');
+import path from 'path';
+import getCallerFile from 'get-caller-file';
 
-module.exports = requireAsHash;
+export default requireAsHash;
 function requireAsHash(pattern, type) {
   let callerFileDir = path.dirname(getCallerFile());
   return globSync(pattern, { cwd: callerFileDir }).reduce((hash, file) => {
